requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1

langchain==0.3.13
langchain-openai==0.2.14
langchain-chroma==0.1.4
chromadb==0.5.23

openai==1.61.0
python-docx==1.1.2
tiktoken==0.8.0